refactor filter_tube_state
We no longer pass the whole database Object but just the required arrays. This way we can always reconstruct this array as it is needed for the views of the reactor. Reason for this change is that it is impossible to store the state for every reactor (array length constraints in postgress)